+++ src/components/Notification/index.js
@@ -1,0 +1,38 @@
+import React, { useState } from "react";
+import { makeStyles } from "@material-ui/core/styles";
+import {
+  Button,
+  Snackbar,
+  IconButton
+} from "@material-ui/core";
+import CloseIcon from '@material-ui/icons/Close';
+
+
+const Notification = ({
+  message, open, handleClose
+}) =>  {
+  {console.log(open)}
+  return (
+    <div>
+      <Snackbar
+        anchorOrigin={{
+          vertical: 'bottom',
+          horizontal: 'left',
+        }}
+        open={open}
+        autoHideDuration={5000}
+        onClose={handleClose}
+        message={message}
+        action={
+          <React.Fragment>
+            <IconButton size="small" aria-label="close" color="inherit" onClick={handleClose}>
+              <CloseIcon fontSize="small" />
+            </IconButton>
+          </React.Fragment>
+        }
+      />
+    </div>
+  );
+}
+
+export { Notification };
