+++ src/components/Ticket/TicketButton.js
@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import {
   Button,
@@ -12,7 +12,6 @@ const useStyles = makeStyles((theme) => ({
     },
   },
 }));
-
 
 const TicketButton = ({
   type, handleClick
